feat(serverless-contracts): remove typebridge
