+++ packages/serverless-contracts/package.json
@@ -45,8 +45,7 @@
     "@babel/runtime": "^7.19.4",
     "ajv": "^8.11.0",
     "http-errors": "^2.0.0",
-    "lodash": "^4.17.21",
-    "typebridge": "^1.2.0"
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
     "@babel/cli": "^7.19.3",
