--- packages/serverless-contracts/src/contracts/event/features/index.ts
@@ -1,1 +1,0 @@
-export { getEventBridgeLambdaTrigger } from './getEventBridgeLambdaTrigger';
