--- packages/serverless-contracts/src/contracts/event/index.ts
@@ -1,3 +1,0 @@
-export { EventContract } from './eventContract';
-export { EventBridgeEventContract } from './eventBridgeEventContract';
-export * from './features';
