--- packages/serverless-contracts/src/contracts/event/types/index.ts
@@ -1,1 +1,0 @@
-export type { EventBridgeLambdaCompleteTriggerType } from './lambdaTrigger';
