+++ packages/serverless-contracts/src/contracts/eventBridge/eventBridgeContract.ts
@@ -1,0 +1,9 @@
+export class EventBridgeContract {
+  public source: string;
+  public name: string;
+
+  constructor({ source, name }: { source: string; name: string }) {
+    this.source = source;
+    this.name = name;
+  }
+}
