+++ packages/serverless-contracts/src/contracts/eventBridge/index.ts
@@ -1,0 +1,1 @@
+export { EventBridgeContract } from './eventBridgeContract';
