+++ packages/serverless-contracts/src/contracts/index.ts
@@ -1,3 +1,3 @@
 export * from './apiGateway';
 export * from './cloudFormation';
-export * from './event';
+export * from './eventBridge';
